playlist/requirements.txt
autopep8==1.6.0
Django==4.1.3
django-taggit==3.1.0
openpyxl==3.0.9
pafy==0.5.5
virtualenv==20.14.1
xkit==0.0.0
xlwt==1.3.0
youtube-dl==2021.12.17
zipp==3.5.0
django-allauth==0.52.0
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
black==22.12.0
celery==5.2.7
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-summernote==0.8.20.0
regex==2021.7.6
pre-commit==2.21.0
